[Serde] Rename encrypt serde to aesgcm; split out key_provider
Rename the L2 encryption serde to consistently use "aesgcm" throughout:
registry type "encrypt" -> "aesgcm", classes EncryptSerializer/
DecryptDeserializer -> AesGcmSerializer/AesGcmDeserializer, module
encrypt.py -> aesgcm.py, and all tests, docs, and the example aligned.

Extract KeyProvider and HkdfKeyProvider into a separate key_provider.py
module so the key-management abstraction is decoupled from the serde.
